+++ spec/source_spec.rb
@@ -2,7 +2,7 @@ require 'spec_helper'
 
 describe 'nagios::default' do
   cached(:chef_run) do
-    ChefSpec::ServerRunner.new(platform: 'ubuntu', version: '16.04') do |node, server|
+    ChefSpec::ServerRunner.new(platform: 'ubuntu', version: '20.04') do |node, server|
       node.normal['nagios']['server']['install_method'] = 'source'
       server.create_data_bag('users',
                                         'user1' => {
@@ -33,11 +33,11 @@ describe 'nagios::default' do
   end
 
   it 'should install the php-gd package' do
-    expect(chef_run).to install_package('php7.0-gd')
+    expect(chef_run).to install_package('php7.4-gd')
   end
 
   it 'should include source install dependency packages' do
-    expect(chef_run).to install_package(%w(libssl-dev libgd2-xpm-dev bsd-mailx tar unzip))
+    expect(chef_run).to install_package(%w(libssl-dev libgdchart-gd2-xpm-dev bsd-mailx tar unzip))
   end
 
   it 'should create nagios user and group' do